Move out tool baseline to ai dir
