+++ src/main/ai/BrowserTool.ts
