+++ src/main/ai/tools/CreateTabTool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { BrowserTool, ok, type ToolResult } from "./BrowserTool";
-import type { ToolContext } from "./ToolContext";
+import { BrowserTool, ok, type ToolResult } from "../BrowserTool";
+import type { ToolContext } from "../ToolContext";
 import { Utils } from "../../utils";
 
 const createTabInputSchema = z.object({
