+++ src/main/ai/tools/ForwardTool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { BrowserTool, ok, fail, type ToolResult } from "./BrowserTool";
-import type { ToolContext } from "./ToolContext";
+import { BrowserTool, ok, fail, type ToolResult } from "../BrowserTool";
+import type { ToolContext } from "../ToolContext";
 
 const forwardInputSchema = z.object({});
