+++ src/main/ai/tools/TypeTool.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { locateElement, typeIntoElement } from "../../page/actions";
-import { BrowserTool, ok, fail, type ToolResult } from "./BrowserTool";
-import type { ToolContext } from "./ToolContext";
+import { BrowserTool, ok, fail, type ToolResult } from "../BrowserTool";
+import type { ToolContext } from "../ToolContext";
 
 const typeInputSchema = z.object({
   index: z.number(),
